Factored MLP into a separate file.
