+++ atmorep/transformer/transformer.py
@@ -15,11 +15,11 @@
 ####################################################################################################
 
 import torch
-import numpy as np
-import math
 
-from atmorep.transformer.transformer_base import MLP, prepare_token
+from atmorep.transformer.mlp import MLP
+from atmorep.transformer.transformer_base import prepare_token
 from atmorep.transformer.transformer_attention import MultiSelfAttentionHead
+
 
 class Transformer(torch.nn.Module) :
