+++ atmorep/transformer/transformer_attention.py
@@ -15,12 +15,8 @@
 ####################################################################################################
 
 import torch
-import numpy as np
-import math
 from enum import Enum
-import code
 
-from atmorep.transformer.axial_attention import AxialAttention
 from atmorep.utils.utils import identity
 from atmorep.transformer.transformer_base import checkpoint_wrapper
